+++ src/core/BrightstarDB/Configuration.cs
@@ -123,11 +123,8 @@ namespace BrightstarDB
             StatsUpdateTimespan = GetApplicationSetting(StatsUpdateTimeSpanName, 0);
 
             // Advanced embedded application settings - read from the brightstar section of the app/web.config
-            var advancedConfiguration = ConfigurationManager.GetSection("brightstar") as BrightstarConfiguration;
-            if (advancedConfiguration != null)
-            {
-                PreloadConfiguration = advancedConfiguration.PreloadConfiguration;
-            }
+            EmbeddedServiceConfiguration = ConfigurationManager.GetSection("brightstar") as EmbeddedServiceConfiguration ??
+                                           new EmbeddedServiceConfiguration();
 
 #endif
 #if !PORTABLE
@@ -281,9 +278,9 @@ namespace BrightstarDB
         public static int StatsUpdateTimespan { get; set; }
 
         /// <summary>
-        /// Get or set the configuration for the page cache warmup
+        /// Get or set the additional configuration options for running an embedded service
         /// </summary>
-        public static PageCachePreloadConfiguration PreloadConfiguration { get; set; }
+        public static EmbeddedServiceConfiguration EmbeddedServiceConfiguration { get; set; }
 
         /// <summary>
         /// Set this property to true to enable the use of virtual nodes in SPARQL