+++ src/core/BrightstarDB/Storage/IStoreManager.cs
@@ -6,15 +6,18 @@ namespace BrightstarDB.Storage
     internal interface IStoreManager
     {
         IEnumerable<string> ListStores(string baseLocation);
-        IStore CreateStore(string storeLocation, bool readOnly = false);
+
+        IStore CreateStore(string storeLocation, bool readOnly = false, bool withTransactionLog = true);
+
         /// <summary>
         /// Create a new store with a specific persistence type
         /// </summary>
         /// <param name="storeLocation">The path to the directory that will contain the store</param>
         /// <param name="persistenceType">The persistence type to use for the main store indexes</param>
         /// <param name="readOnly">Specifies whether the created store should be returned open for read/write or read only</param>
+        /// <param name="withTransactionLog">Specifies whether the created store should have a transaction log file created for it</param>
         /// <returns>The newly created store, opened for read or read/write depending on the setting of the <paramref name="readOnly"/> parameter</returns>
-        IStore CreateStore(string storeLocation, PersistenceType persistenceType, bool readOnly = false);
+        IStore CreateStore(string storeLocation, PersistenceType persistenceType, bool readOnly = false, bool withTransactionLog = true);
 
         /// <summary>
         /// Creates a new store with the default persistence type